Use the client's 'unique tag' as identifier for requested semaphores
Previously, using just the client member's 'unique Id' did not work as,
when clients of different versions connect to a server, the uniqueId on
the server may also contain a version string. Thus the Id sent by the
client (and used to track a semaphore) might be different to what the
server sees in the DistributedMember.getUniqueId. The latter is used to
clean up semaphores when a client disconnects. With this mismatch,
semaphores would not get cleaned up and released.